refactor: app wiring; upgrade dependency; fix ap news router formatting issue; remove cached items TTL
