+++ application/router_dependencies.py
@@ -1,0 +1,37 @@
+from router.embassy.china_embassy_news_constants import china_embassy_news_filter
+from router.meta_blog.meta_tech_blog_router_constants import meta_blog_prefix
+from router.zaobao.zaobao_realtime_router_constants import zaobao_region_parameter, title_filter
+from application.router_objects import (
+    apnews_business,
+    apnews_top_news,
+    apple_developer_news,
+    apple_newsroom,
+    chinese_embassy_news,
+    cnbeta,
+    day_one_blog,
+    jandan_news,
+    meta_tech_blog,
+    reuters_news,
+    sony_alpha_rumors,
+    usgs_earthquake_report,
+    wsdot_news,
+    yahoo_news,
+    zaobao_realtime,
+)
+from utils.router_constants import (
+    apnews_business_router_path,
+    apnews_router_path,
+    apple_news_router_path,
+    apple_newsroom_router_path,
+    cnbeta_router_path,
+    day_one_blog_router_path,
+    earthquake_router_path,
+    embassy_router_path,
+    jandan_router_path,
+    meta_engineering_blog_router as meta_engineering_blog_router_path,
+    reuters_news_router_path,
+    sar_router_path,
+    wsdot_news_router_path,
+    yahoo_news_router_path_prefix,
+    zaobao_router_path_prefix,
+)
