+++ requirements.txt
@@ -5,8 +5,8 @@ certifi==2026.4.22
 charset-normalizer==3.4.7
 feedparser==6.0.12
 Flask==3.1.3
-idna==3.13
-importlib-metadata==8.7.0
+idna==3.14
+importlib-metadata==9.0.0
 itsdangerous==2.2.0
 Jinja2==3.1.6
 lxml==5.4.0
@@ -15,7 +15,7 @@ python-dateutil==2.9.0.post0
 PyJWT==2.12.1
 pytz==2026.2
 PyYAML==6.0.3
-requests==2.33.1
+requests==2.34.0
 rfeed==1.1.1
 sgmllib3k==1.0.0
 redis==5.3.1
@@ -23,6 +23,6 @@ six==1.17.0
 soupsieve==2.8.3
 typing_extensions==4.15.0
 tzlocal==5.3.1
-urllib3==2.6.3
+urllib3==2.7.0
 Werkzeug==3.1.8
 zipp==3.23.1
