+++ utils/router_constants.py
@@ -15,7 +15,6 @@ apple_news_router_path = "/apple/developer"
 apple_newsroom_router_path = "/apple/newsroom"
 
 refresh_period_in_minutes = 15
-metadata_list_max_items = 500
 
 html_parser = 'html.parser'
 
@@ -24,7 +23,6 @@ language_english = "en-us"
 
 # Cache TTL defaults (in seconds)
 last_build_time_ttl_seconds = 3600      # 1 hour
-feed_item_ttl_seconds = 604800          # 7 days
 
 # Distributed lock TTL for warm_cache (in seconds)
 warm_lock_ttl_seconds = 300             # 5 minutes
